🔨 Refactor: Grade 컴포넌트 props 제거
